Move test_macros to a separate file
